fix default html rendering
